requirements.txt
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
httplib2==0.31.0
idna==3.10
joblib==1.5.2
Levenshtein==0.27.1
numpy==2.3.4
pandas==2.3.3
praw==7.8.1
prawcore==2.4.0
proto-plus==1.26.1
protobuf==6.33.0
psaw==0.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
RapidFuzz==3.14.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
soupsieve==2.8
thefuzz==0.22.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
